linkification now catches unicode characters, like japanese
The regex now just grabs all characters that aren't whitespace
or a < (opening <br>). Technically, such characters are
supposed to be encoded with % escapes, but the /a/ daily
japanese thread doesn't seem to do that, probably because
their browser automatically replaced the % escapes from the
URL bar with the actual characters they represent.