Align DOM parser code
Tabularlize is pretty nice.

As to the actual usefullness of space-aligning text, I admit that
doesn't really highlight any *useful* properties of the code, and it
makes it harder to make changes by hand, and it's another one of those
things that we wouldn't have to worry about if elastic tabs were
actually a thing, but hey, it does look pretty nice.