Remove global expansion/preloading
Killin' features like I'm a gnome 3 dev...

The uses for preloading/expansion are too small to warrant their extra ui all the time.

Inline-expansion for images now cycles to 100% from fit-width if the
image is wider than the page, which is pretty okay in almost all cases.

Would be still nice to have a 'manga dump' mode, but I'm not sure how
the UI would best look. Likebox-style left-right to images that match
the same user/size/regex pattern on filename? It would be a lot code for
a tiny feature though.