Unlicense html5chan
After much deliberation, I've decided that copyrights are a solution to
a problem that shouldn't exist. Though I was originally leaning towards
the GPL, because I support the idea that all creative output _should_ be
free and that copyleft is an okay way to nudge people towards that goal,
the requirement to include a license larger than the SLOC of html5chan
(+file headers) was just too... tl;dr, I guess. All that bookkeeping,
making sure I comply with 14 sections of legalese... just too much for
a little 4chan project.

I also like the Unlicense because the notion of disclaiming copyright
altogether appeals to me. While I still don't like the idea that
somebody could (potentially) get famous using work that I did, I
see that more as _my_ failure not to publish and publicize my work and
establish original authorship. If the work is good enough to make
somebody famous without significant modification, than it's my own
fault for not letting people know I made it when it becomes public.

I thought about the WTFPL too, but I think the philosophy of Unlicense
is stronger. Plus, it's easier to mention in mixed company.