Require browserify from Cokefile
I was going to actually implement the compiling shit, but apparently
browserify just changed its entire API, which is good in that the
compiled output is _way_ cleaner, but bad in that now I have to figure
out how to get it to compile coco. From a cursory glance at the source,
it's got a whole bunch of weird stream/socket shit, and comments are
kinda sparse. Not that I don't like streams (bacon.js), but it's like
reading haskell--I know the syntax, but the libraries are so foreign to
me that it's impossible to tell what's going on.

So I'll just push this as a topic branch and finish this... later