Fix backlinks not being added to inserted posts
The backlinks concern is way ugly now. Where did I go wrong?

One alternative is to adopt mayhem's 4chanx v3's clone thing, but that
still has undesirable amounts of state associated with it.

Cleaner backlinks is probably more dependent on a cleaner thread updater
first, since if backlinks never changed, they would be a lot simpler, so
the complexity of one is tied to the other.