Add basic unread post tracking
Kind of ugly css, since becoming read shifts posts to the left by 5px.

Also not a great implementation, since posts barely scrolled over will
be marked as read, as well as posts that are above the top of the page
if a specific post hash is used.

Will fix later.