Improve catalog page display
With header and css this time.

Using -moz-columns to give an intstagram-like layout without using
masonry or isotope or whatever is popular these days. Probably will want
to use something like that in the end, because columns are annoying to
work with.

Tried to get the teaser text to expand on hover without affecting the
layout, but my CSS wizardry isn't up to snuff lately. Will try again
later.

All these new lines of code... my SLOC ;_;