Catalog: more intelligently sort updated
It looks a bit weird when you actually see the numbers
(why is this thread at the top when this thread has
so many more new replies?). Might need some tweaking
to promote "healthy" threads with 10+ replies over
threads with 1 new reply / 1 total (top of the list
in this sort).